refactor: изменить тип полей id с Long на Integer
Поможет избежать излишнего приведения типов

+++ src/main/java/ru/practicum/shareit/item/model/Item.java
@@ -14,10 +14,10 @@ import javax.validation.constraints.Size;
  * <p><strong>Вещь</strong> - основная сущность сервиса, вокруг которой строится вся работа.</p>
  * <p>Data-класс Item содержит поля:</p>
  * <ul>
- *     <li>id - уникальный идентификатор вещи, тип Long;</li>
+ *     <li>id - уникальный идентификатор вещи, тип Integer;</li>
  *     <li>name - краткое название, тип String;</li>
  *     <li>description - развёрнутое описание, тип String;</li>
- *     <li>ownerId - уникальный идентификатор владельца вещи, тип Long</li>
+ *     <li>ownerId - уникальный идентификатор владельца вещи, тип Integer</li>
  *     <li>available - статус о доступности вещи для аренды, тип Boolean.</li>
  * </ul>
  */
@@ -26,14 +26,14 @@ import javax.validation.constraints.Size;
 @Builder
 @AllArgsConstructor
 public class Item {
-    @Min(value = 1L, message = "Идентификатор должен быть положительным числом")
-    private Long id;
+    @Min(value = 1, message = "Идентификатор должен быть положительным числом")
+    private Integer id;
     @NotBlank(message = "Название не может быть пустым")
     private String name;
     @Size(max = 255, message = "Размер описания не должен превышать 255 символов")
     private String description;
     @NotNull(message = "У вещи должен быть владелец")
-    @Min(value = 1L, message = "Идентификатор должен быть положительным числом")
-    private Long ownerId;
+    @Min(value = 1, message = "Идентификатор должен быть положительным числом")
+    private Integer ownerId;
     private Boolean available;
 }
