+++ src/main/java/ru/practicum/shareit/user/model/User.java
@@ -11,7 +11,7 @@ import javax.validation.constraints.NotBlank;
 /**
  * <p>Data-класс User содержит информацию о сущности "пользователь" и имеет поля:</p>
  * <ul>
- *     <li>id - уникальный идентификатор пользователя, тип Long;</li>
+ *     <li>id - уникальный идентификатор пользователя, тип Integer;</li>
  *     <li>name - имя пользователя, тип String;</li>
  *     <li>email - адрес электронной почты пользователя, тип String.</li>
  * </ul>
@@ -21,8 +21,8 @@ import javax.validation.constraints.NotBlank;
 @Builder
 @AllArgsConstructor
 public class User {
-    @Min(value = 1L, message = "Идентификатор должен быть положительным числом")
-    private Long id;
+    @Min(value = 1, message = "Идентификатор должен быть положительным числом")
+    private Integer id;
     private String name;
     @NotBlank
     @Email(message = "Некорректный email")
